test : 랭크 개수 카운트 테스트코드 작성
